launcher model branched into macos and linux
